Fix race condition on direction arrays
The function to find a random receiver uses a global array which
dictates the legal directions in a given topology. Randomness
is achieved by shuffling this array.
If concurrent calls overlap, multiple shuffling can take place,
preventing to check some regions. This can generate errors in
corner regions and makes the selection not really uniformly random.
The fix is to make a local copy of the array before shuffling it.

Signed-off-by: Alessandro Pellegrini <a.pellegrini@ing.uniroma2.it>